feat: expose tool functions via AI helper
